Use a separate class for global settings
