+++ src/Datadog.Trace/Configuration/GlobalSettings.cs
@@ -1,0 +1,128 @@
+using System.IO;
+using Datadog.Trace.Logging;
+using Datadog.Trace.Vendors.Serilog.Events;
+
+namespace Datadog.Trace.Configuration
+{
+    /// <summary>
+    /// Contains global datadog settings.
+    /// </summary>
+    public class GlobalSettings
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="GlobalSettings"/> class with default values.
+        /// </summary>
+        public GlobalSettings()
+            : this(null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="GlobalSettings"/> class
+        /// using the specified <see cref="IConfigurationSource"/> to initialize values.
+        /// </summary>
+        /// <param name="source">The <see cref="IConfigurationSource"/> to use when retrieving configuration values.</param>
+        public GlobalSettings(IConfigurationSource source)
+        {
+            DebugEnabled = source?.GetBool(ConfigurationKeys.DebugEnabled) ??
+                           // default value
+                           false;
+
+            LogsInjectionEnabled = source?.GetBool(ConfigurationKeys.LogsInjectionEnabled) ??
+                                   // default value
+                                   false;
+        }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether debug mode is enabled.
+        /// Default is <c>false</c>.
+        /// </summary>
+        /// <seealso cref="ConfigurationKeys.DebugEnabled"/>
+        public bool DebugEnabled { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether correlation identifiers are
+        /// automatically injected into the logging context.
+        /// Default is <c>false</c>.
+        /// </summary>
+        /// <seealso cref="ConfigurationKeys.LogsInjectionEnabled"/>
+        public bool LogsInjectionEnabled { get; set; }
+
+        /// <summary>
+        /// Gets the global settings instance.
+        /// </summary>
+        internal static GlobalSettings Source { get; } = FromDefaultSources();
+
+        /// <summary>
+        /// Set whether debug mode is enabled.
+        /// Affects the level of logs written to file.
+        /// </summary>
+        /// <param name="enabled">Whether debug is enabled.</param>
+        public static void SetDebugEnabled(bool enabled)
+        {
+            Source.DebugEnabled = enabled;
+
+            if (enabled)
+            {
+                DatadogLogging.SetLogLevel(LogEventLevel.Verbose);
+            }
+            else
+            {
+                DatadogLogging.UseDefaultLevel();
+            }
+        }
+
+        /// <summary>
+        /// Create a <see cref="GlobalSettings"/> populated from the default sources
+        /// returned by <see cref="CreateDefaultConfigurationSource"/>.
+        /// </summary>
+        /// <returns>A <see cref="TracerSettings"/> populated from the default sources.</returns>
+        public static GlobalSettings FromDefaultSources()
+        {
+            var source = CreateDefaultConfigurationSource();
+            return new GlobalSettings(source);
+        }
+
+        /// <summary>
+        /// Creates a <see cref="IConfigurationSource"/> by combining environment variables,
+        /// AppSettings where available, and a local datadog.json file, if present.
+        /// </summary>
+        /// <returns>A new <see cref="IConfigurationSource"/> instance.</returns>
+        internal static CompositeConfigurationSource CreateDefaultConfigurationSource()
+        {
+            // env > AppSettings > datadog.json
+            var configurationSource = new CompositeConfigurationSource
+            {
+                new EnvironmentConfigurationSource(),
+
+#if !NETSTANDARD2_0
+                // on .NET Framework only, also read from app.config/web.config
+                new NameValueConfigurationSource(System.Configuration.ConfigurationManager.AppSettings)
+#endif
+            };
+
+            string currentDirectory = System.Environment.CurrentDirectory;
+
+#if !NETSTANDARD2_0
+            // on .NET Framework only, use application's root folder
+            // as default path when looking for datadog.json
+            if (System.Web.Hosting.HostingEnvironment.IsHosted)
+            {
+                currentDirectory = System.Web.Hosting.HostingEnvironment.MapPath("~");
+            }
+#endif
+
+            // if environment variable is not set, look for default file name in the current directory
+            var configurationFileName = configurationSource.GetString(ConfigurationKeys.ConfigurationFileName) ??
+                                        Path.Combine(currentDirectory, "datadog.json");
+
+            if (Path.GetExtension(configurationFileName).ToUpperInvariant() == ".JSON" &&
+                File.Exists(configurationFileName))
+            {
+                configurationSource.Add(JsonConfigurationSource.FromFile(configurationFileName));
+            }
+
+            return configurationSource;
+        }
+    }
+}
