+++ src/Datadog.Trace/Logging/DatadogLogging.cs
@@ -27,7 +27,7 @@ namespace Datadog.Trace.Logging
                    .CreateLogger();
             try
             {
-                if (TracerSettings.GlobalSettings.DebugEnabled)
+                if (GlobalSettings.Source.DebugEnabled)
                 {
                     LoggingLevelSwitch.MinimumLevel = LogEventLevel.Verbose;
                 }
